requirements.txt
fastapi
sqlalchemy
uvicorn
psycopg2-binary
cryptography
PyJWT 
requests
opentelemetry-exporter-otlp
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
# psycopg2
